fix(unit tests): moved to separate unit test directory
