+++ test/unit/a08-validators.spec.js
@@ -1,7 +1,7 @@
 const assert = require('chai').assert
 const testUtils = require('./utils')
 
-const Validators = require('../src/middleware/validators')
+const Validators = require('../../src/middleware/validators')
 
 const sinon = require('sinon')
 const mockContext = require('./mocks/ctx-mock').context
