+++ test/unit/a09-admin.spec.js
@@ -1,6 +1,6 @@
 const assert = require('chai').assert
 
-const Admin = require('../src/lib/admin')
+const Admin = require('../../src/lib/admin')
 
 const sinon = require('sinon')
