+++ test/unit/mocks/log-api-mock.js
