+++ test/unit/utils.js
@@ -1,5 +1,5 @@
 const mongoose = require('mongoose')
-const config = require('../config')
+const config = require('../../config')
 const axios = require('axios').default
 
 const LOCALHOST = `http://localhost:${config.port}`
@@ -79,7 +79,7 @@ async function loginTestUser () {
 
 async function loginAdminUser () {
   try {
-    const FILENAME = `../config/system-user-${config.env}.json`
+    const FILENAME = `../../config/system-user-${config.env}.json`
     const adminUserData = require(FILENAME)
     console.log(`adminUserData: ${JSON.stringify(adminUserData, null, 2)}`)
 
@@ -115,7 +115,7 @@ async function getAdminJWT () {
     // process.env.KOA_ENV = process.env.KOA_ENV || 'dev'
     // console.log(`env: ${process.env.KOA_ENV}`)
 
-    const FILENAME = `../config/system-user-${config.env}.json`
+    const FILENAME = `../../config/system-user-${config.env}.json`
     const adminUserData = require(FILENAME)
     // console.log(`adminUserData: ${JSON.stringify(adminUserData, null, 2)}`)
